experiments/requirements.txt
hydra-core
ujson
jsonargparse
lightning
openai
tenacity
evaluate
transformers
scikit-learn
